restruct the files
